requirements.txt
streamlit
pandas
numpy
altair
umap-learn
scikit-learn